Switched to minute based ephemeris by JPL. Fails.
In fact, it seems like the post conversion spacecraft is still just as
far away as the VSOP87 one. However, many tests have showns that the JPL
data was quite different from the VSOP87. So I'll stick with the JPL
information for now and fix the reference frame change.

Also fixed dates in astro.go to be solely UTC. Avoids issues when
looking up Ephemeris.

Check Thesis list on Google Keep. I found the answer in Dr. Schaub's
book.