Ruggiero seems to work but...
One of the main problems is that it assigns the same weights to each OE,
and that leads to the smaller ones being more important than others. For
example, initially, the semi major axis change is the biggest, but it
quickly becomes very small. In the last 8-9 months of a basic GEO
propagation, the inclination is 0.019 (instead of 0.0), which means that
the thrust vector has a 0.015 weight on that. But the semi major axis is
at 42245km instead of 42164, which yields a factor of -0.00460, so
completely negligible in comparasion. But that semi major axis change is
HUGE, so it should be fixed first. Gotta find a better strategy here.