Making opti only on semi-major axis converges
However, it pursues the target until that reaches. I think I need to
have an additional stop conditions which is on the distance from the Sun
and more specifically the proximity to the *theoritical* SOI. Then, it
should do an orbital injection (which currently actually fails, probably
because the planet isn't there). Then gotta change opti target to focus
on several orbital elements and also stop at a given distance to the
target.