Something wrong with CL of inclination
Seems to work if I am always applying the maximum thrust. The weird
thing is that this law uses the argument of perigee and the true
anomaly, I've checked many many times that all of the dx/dt functions
are correct. So, I'm confused, quite seriously here. Wait, actually, the
d(true anomaly)/dt comes from Vallado, and I haven't double checked that
one with other sources, I only check it against the GEO propagation. In
addition, it's effectively the same law for the RAAN change, but with
math.Sin instead of math.Cos, so if the true anomaly is wrong, it makes
sense for the RAAN to also be wrong (i.e. not work).