Improving coverage. Big issue: orbit prop is wrong
In fact, propagating on time only does not actually change anything.
Only the time is added, but the true anomaly does not change. This
somewhat makes sense because the thrusting vector is nil, but is should
still do something I think because the true anomaly does change with
time... Not sure what is going on there.