Added control laws in OrbitTarget
Had to update the signation to NewOptimalΔOrbit.
TestMultiCorrectOE leads to a panic. I suspect it's due to the
eccentricity correction, but I'm not sure yet. If not, it's due to the
Ruggerio factor.

Note: the inclination change does not work at all with the provided
equation, regardless of the duration of propagation (I tried a 20 times
longer propagation).