Some Naasz tests fail but are very close, weird.
For example, the argument of perigee test is with 0.016 instead of 0.01
degrees from the passing value. Same goes for the RAAN change (I think).
Similarly, decreasing the semi major axis is about 300km off, although
the change is a 18Mm change, so it's damn close. Extending the
propagation time for the aNeg test does not help (nor does making the
distance epsilon smaller), so this is just weird. It's even weirder that
these are *single* OE tests, nothing more.