YEAH BABYYYY! Continuous works =D
Targetting the planet using Naasz and only a, i, e leads to reaching
that orbit. However, unsurprisingly, the orbit is hyperbolic so the
propagator kills it at the next time step effectively acting like a
flyby. This is a great start! Now on to implementing a nice spiral so
that I can spiral out of Mars and target that orbit on the SOI border.