The injectino does not work. *But* I have an idea.
I think it fails because I'm using SPICE to do the conversion and when the spacecraft gets to the planet, Mars isn't actually there. So I need to do a good targeting. Otherwise, it's evident that I cannot perform any injection around Mars.
Secondary idea: use/code a differential corrector for the initial leaving time for proper targetting? Also since it's only in 1D, it's actually just a simple dichotomy like algorithm