MAJOR ISSUE: Phi in Estimator is that from time 0 (not time k-1).
This makes the smoother very angry. I have implemented a fix to return the correct Phi, but that leads to the KFs being broken. They break as soon as there is a large time gap between measurements. This is because I'm doing all the plots in measurement and not in time. The only good news here is that I know that my new Phis are correct (compared with Ben's). This will require several long hours of work to fix.