Fixed Ruggiero to work with small differences in initial and final OEs
Until now it would lead to a NaN error, so it was prevented. however,
that causes Case A of Petropoulos to fail because the eccentricity was
never correct on with Ruggerio. In fact, the sma correction led to a
change in ecc, but because the initial and target ecc was equal, no
correction could mathematically be applied without causing a NaN.